Move TestHelper module into different modules
